refactor and usage of service discount

+++ CleanCode/Services/DiscountService.cs
@@ -4,10 +4,10 @@ namespace CleanCode.Services
     {
         private const double PERCENT_100 = 100;
 
-        public decimal GetDiscountedPrice(decimal price, decimal discountInPercent) =>
+        public decimal GetDiscountedPrice(decimal price, double discountInPercent) =>
             price - GetDiscount(price, discountInPercent);
 
-        private static decimal GetDiscount(decimal price, decimal discountInPercent) =>
-            price * discountInPercent / (decimal) PERCENT_100;
+        private static decimal GetDiscount(decimal price, double discountInPercent) =>
+            (decimal)((double)price * discountInPercent / PERCENT_100);
     }
 }
