+++ CleanCode/Services/IDiscountService.cs
@@ -2,6 +2,6 @@ namespace CleanCode.Services
 {
     public interface IDiscountService
     {
-        public decimal GetDiscountedPrice(decimal price, decimal discountInPercent);
+        public decimal GetDiscountedPrice(decimal price, double discountInPercent);
     }
 }
